requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
cycler==0.10.0
decorator==4.4.2
fsspec==2021.7.0
future==0.18.2
google-auth==1.33.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
idna==3.2
imageio==2.9.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
multidict==5.1.0
networkx==2.5.1
numpy==1.21.1
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.3.8
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.7.2
torch==1.9.0
torchmetrics==0.4.1
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
